Python/Networking/Messages Network/Test Unit/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\helpers\helper_functions.py: 9
PyGithub == 1.55

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\UI\chat_tab.py: 1,2
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\UI\main_window.py: 1,2
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\client\client.py: 7
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\main.py: 4,5,6
PyQt5 == 5.15.6

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\helpers\helper_functions.py: 13
mss == 6.1.0

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\helpers\helper_functions.py: 14
numpy == 1.21.2

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\client\client.py: 4
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\helpers\constants.py: 1
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\helpers\helper_functions.py: 12
# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\server\server.py: 7
opencv_python == 4.5.4.58

# C:\Users\Personne\Desktop\Coding Projects\Python\Networking\Messages Network\Test Unit\main.py: 14
requests == 2.26.0
